+++ imperial-red/src/main/java/urlshortener2015/imperialred/repository/ShortURLRepositoryCustom.java
@@ -1,12 +1,13 @@
 package urlshortener2015.imperialred.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import urlshortener2015.imperialred.objects.ShortURL;
 
 public interface ShortURLRepositoryCustom {
 	
-	List<ShortURL> list(Long limit, Long offset);
+	List<ShortURL> list(BigInteger limit, BigInteger offset);
 	
 	void update(ShortURL su);
